break out ProviderMachineSerializer
